requirements.txt
appdirs==1.4.0
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
certifi==2016.2.28
chainer==1.8.1
chardet==2.3.0
decorator==4.0.9
filelock==2.0.5
igo-python==0.9.7
mecab-python==0.996
nose==1.3.7
numpy==1.11.0
protobuf==2.6.1
py==1.4.31
pytest==2.9.1
pytools==2016.1
singledispatch==3.4.0.3
six==1.10.0
tornado==4.3
virtualenv==15.0.1
